Bug 1198458: Webrtc updated to branch 43; pull made 2015-09-29 09:00AM PDT rs=jesup https://chromium.googlesource.com/external/webrtc git checkout -b my_branch refs/remotes/branch-heads/43 commit 6979b024d7cebfdcd1e8f66da59ea157bbc9e47e (I believe)

+++ media/webrtc/trunk/webrtc/modules/audio_coding/codecs/cng/include/webrtc_cng.h
@@ -12,6 +12,7 @@
 #ifndef WEBRTC_MODULES_AUDIO_CODING_CODECS_CNG_MAIN_INTERFACE_WEBRTC_CNG_H_
 #define WEBRTC_MODULES_AUDIO_CODING_CODECS_CNG_MAIN_INTERFACE_WEBRTC_CNG_H_
 
+#include <stddef.h>
 #include "webrtc/typedefs.h"
 
 #ifdef __cplusplus
@@ -120,7 +121,7 @@ int16_t WebRtcCng_Encode(CNG_enc_inst* cng_inst, int16_t* speech,
  *                      -1 - Error
  */
 int16_t WebRtcCng_UpdateSid(CNG_dec_inst* cng_inst, uint8_t* SID,
-                            int16_t length);
+                            size_t length);
 
 /****************************************************************************
  * WebRtcCng_Generate(...)
