+++ media/webrtc/trunk/webrtc/modules/audio_conference_mixer/source/time_scheduler.h
@@ -22,7 +22,7 @@ class CriticalSectionWrapper;
 class TimeScheduler
 {
 public:
-    TimeScheduler(const uint32_t periodicityInMs);
+    TimeScheduler(const int64_t periodicityInMs);
     ~TimeScheduler();
 
     // Signal that a periodic event has been triggered.
@@ -30,7 +30,7 @@ public:
 
     // Set updateTimeInMs to the amount of time until UpdateScheduler() should
     // be called. This time will never be negative.
-    int32_t TimeToNextUpdate(int32_t& updateTimeInMS) const;
+    int32_t TimeToNextUpdate(int64_t& updateTimeInMS) const;
 
 private:
     CriticalSectionWrapper* _crit;
@@ -38,8 +38,8 @@ private:
     bool _isStarted;
     TickTime _lastPeriodMark;
 
-    uint32_t _periodicityInMs;
-    int64_t  _periodicityInTicks;
+    int64_t _periodicityInMs;
+    int64_t _periodicityInTicks;
     uint32_t _missedPeriods;
 };
 }  // namespace webrtc
