+++ media/webrtc/trunk/webrtc/modules/video_coding/main/source/session_info.h
@@ -22,7 +22,7 @@ namespace webrtc {
 // Used to pass data from jitter buffer to session info.
 // This data is then used in determining whether a frame is decodable.
 struct FrameData {
-  int rtt_ms;
+  int64_t rtt_ms;
   float rolling_average_packets_per_frame;
 };
 
@@ -56,15 +56,15 @@ class VCMSessionInfo {
   // Builds fragmentation headers for VP8, each fragment being a decodable
   // VP8 partition. Returns the total number of bytes which are decodable. Is
   // used instead of MakeDecodable for VP8.
-  int BuildVP8FragmentationHeader(uint8_t* frame_buffer,
-                                  int frame_buffer_length,
-                                  RTPFragmentationHeader* fragmentation);
+  size_t BuildVP8FragmentationHeader(uint8_t* frame_buffer,
+                                     size_t frame_buffer_length,
+                                     RTPFragmentationHeader* fragmentation);
 
   // Makes the frame decodable. I.e., only contain decodable NALUs. All
   // non-decodable NALUs will be deleted and packets will be moved to in
   // memory to remove any empty space.
   // Returns the number of bytes deleted from the session.
-  int MakeDecodable();
+  size_t MakeDecodable();
 
   // Sets decodable_ to false.
   // Used by the dual decoder. After the mode is changed to kNoErrors from
@@ -72,7 +72,7 @@ class VCMSessionInfo {
   // decodable and are not complete are marked as non-decodable.
   void SetNotDecodableIfIncomplete();
 
-  int SessionLength() const;
+  size_t SessionLength() const;
   int NumPackets() const;
   bool HaveFirstPacket() const;
   bool HaveLastPacket() const;
@@ -114,8 +114,8 @@ class VCMSessionInfo {
   PacketIterator FindPartitionEnd(PacketIterator it) const;
   static bool InSequence(const PacketIterator& it,
                          const PacketIterator& prev_it);
-  int InsertBuffer(uint8_t* frame_buffer,
-                   PacketIterator packetIterator);
+  size_t InsertBuffer(uint8_t* frame_buffer,
+                      PacketIterator packetIterator);
   size_t Insert(const uint8_t* buffer,
                 size_t length,
                 bool insert_start_code,
@@ -124,8 +124,8 @@ class VCMSessionInfo {
   PacketIterator FindNaluEnd(PacketIterator packet_iter) const;
   // Deletes the data of all packets between |start| and |end|, inclusively.
   // Note that this function doesn't delete the actual packets.
-  int DeletePacketData(PacketIterator start,
-                       PacketIterator end);
+  size_t DeletePacketData(PacketIterator start,
+                          PacketIterator end);
   void UpdateCompleteSession();
 
   // When enabled, determine if session is decodable, i.e. incomplete but
